# public/swagger/v1/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Severity": {
				"enum": [
					"Low",
					"Medium"
				],
				"type": "string"
			},
			"SeverityStatusHistory1": {
				"properties": {
					"severity": {
						"$ref": "#/components/schemas/Severity"
					},
					"date": {
						"type": "number",
						"format": "double"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					}
				},
				"required": [
					"severity",
					"date",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Threat1": {
				"enum": [
					"",
					"Vulnerability"
				],
				"type": "string"
			},
			"ThreatAffect1": {
				"properties": {
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					},
					"platformText": {
						"type": "string"
					}
				},
				"required": [
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf",
					"platformText"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RemediationStep1": {
				"properties": {
					"link": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					},
					"step": {
						"type": "string"
					}
				},
				"required": [
					"link",
					"type",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf",
					"step"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CveIdentifier1": {
				"properties": {
					"cveStatus": {
						"type": "string"
					},
					"cveIdentifier": {
						"type": "string"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					},
					"cveText": {
						"type": "string"
					}
				},
				"required": [
					"cveStatus",
					"cveIdentifier",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf",
					"cveText"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SectionType": {
				"enum": [
					"website-section"
				],
				"type": "string"
			},
			"Section1": {
				"properties": {
					"sectionType": {
						"$ref": "#/components/schemas/SectionType"
					},
					"isNumberedList": {
						"type": "boolean"
					},
					"title": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					},
					"html": {
						"type": "string"
					}
				},
				"required": [
					"sectionType",
					"isNumberedList",
					"title",
					"type",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf",
					"html"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Category": {
				"enum": [
					"",
					"Exploit"
				],
				"type": "string"
			},
			"Item1": {
				"properties": {
					"threatId": {
						"type": "string"
					},
					"sourceOfThreatUpdates": {
						"items": {},
						"type": "array"
					},
					"cyberAcknowledgements": {
						"items": {},
						"type": "array"
					},
					"severityStatusHistory": {
						"items": {
							"$ref": "#/components/schemas/SeverityStatusHistory1"
						},
						"type": "array"
					},
					"threatType": {
						"$ref": "#/components/schemas/Threat1"
					},
					"threatvector": {
						"items": {
							"$ref": "#/components/schemas/Threat1"
						},
						"type": "array"
					},
					"threatAffects": {
						"items": {
							"$ref": "#/components/schemas/ThreatAffect1"
						},
						"type": "array"
					},
					"threatUpdates": {
						"items": {},
						"type": "array"
					},
					"remediationSteps": {
						"items": {
							"$ref": "#/components/schemas/RemediationStep1"
						},
						"type": "array"
					},
					"indicatorsCompromise": {
						"items": {},
						"type": "array"
					},
					"ncscLink": {
						"type": "string"
					},
					"cveIdentifiers": {
						"items": {
							"$ref": "#/components/schemas/CveIdentifier1"
						},
						"type": "array"
					},
					"shortsummary": {
						"type": "string"
					},
					"publishedDate": {
						"type": "number",
						"format": "double"
					},
					"severity": {
						"$ref": "#/components/schemas/Severity"
					},
					"sections": {
						"items": {
							"$ref": "#/components/schemas/Section1"
						},
						"type": "array"
					},
					"basePath": {
						"type": "string"
					},
					"category": {
						"items": {
							"$ref": "#/components/schemas/Category"
						},
						"type": "array"
					},
					"fullTaxonomyList": {
						"items": {},
						"type": "array"
					},
					"title": {
						"type": "string"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"remediationIntro": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					}
				},
				"required": [
					"threatId",
					"sourceOfThreatUpdates",
					"cyberAcknowledgements",
					"severityStatusHistory",
					"threatType",
					"threatvector",
					"threatAffects",
					"threatUpdates",
					"remediationSteps",
					"indicatorsCompromise",
					"ncscLink",
					"cveIdentifiers",
					"shortsummary",
					"publishedDate",
					"severity",
					"sections",
					"basePath",
					"category",
					"fullTaxonomyList",
					"title",
					"versionedNode",
					"remediationIntro",
					"summary"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Welcome": {
				"properties": {
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"items": {
						"items": {
							"$ref": "#/components/schemas/Item1"
						},
						"type": "array"
					},
					"currentPage": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"pageSize",
					"total",
					"items",
					"currentPage",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Section": {
				"properties": {
					"title": {
						"type": "string"
					},
					"isNumberedList": {
						"type": "boolean"
					},
					"headingLevel": {
						"type": "string"
					},
					"sectionType": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					},
					"html": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"isNumberedList",
					"headingLevel",
					"sectionType",
					"type",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf",
					"html"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SeverityStatusHistory": {
				"properties": {
					"severity": {
						"type": "string"
					},
					"date": {
						"type": "number",
						"format": "double"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					}
				},
				"required": [
					"severity",
					"date",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IndicatorsCompromise": {
				"properties": {
					"audience": {
						"type": "string"
					},
					"heading": {
						"type": "string"
					},
					"sectionType": {
						"type": "string"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"audience",
					"heading",
					"sectionType",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CveIdentifier": {
				"properties": {
					"cveIdentifier": {
						"type": "string"
					},
					"cveStatus": {
						"type": "string"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					},
					"cveText": {
						"type": "string"
					}
				},
				"required": [
					"cveIdentifier",
					"cveStatus",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf",
					"cveText"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ThreatAffect": {
				"properties": {
					"versionsAffected": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					},
					"platformText": {
						"type": "string"
					}
				},
				"required": [
					"versionsAffected",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf",
					"platformText"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RemediationStep": {
				"properties": {
					"link": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"hippoDocumentBean": {
						"type": "boolean"
					},
					"hippoFolderBean": {
						"type": "boolean"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"leaf": {
						"type": "boolean"
					},
					"step": {
						"type": "string"
					}
				},
				"required": [
					"link",
					"type",
					"hippoDocumentBean",
					"hippoFolderBean",
					"versionedNode",
					"leaf",
					"step"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Welcome1": {
				"properties": {
					"severity": {
						"type": "string"
					},
					"category": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"basePath": {
						"type": "string"
					},
					"sections": {
						"items": {
							"$ref": "#/components/schemas/Section"
						},
						"type": "array"
					},
					"threatId": {
						"type": "string"
					},
					"shortsummary": {
						"type": "string"
					},
					"publishedDate": {
						"type": "number",
						"format": "double"
					},
					"fullTaxonomyList": {
						"items": {},
						"type": "array"
					},
					"seosummaryJson": {
						"type": "string"
					},
					"sourceOfThreatUpdates": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"cyberAcknowledgements": {
						"items": {},
						"type": "array"
					},
					"severityStatusHistory": {
						"items": {
							"$ref": "#/components/schemas/SeverityStatusHistory"
						},
						"type": "array"
					},
					"indicatorsCompromise": {
						"items": {
							"$ref": "#/components/schemas/IndicatorsCompromise"
						},
						"type": "array"
					},
					"cveIdentifiers": {
						"items": {
							"$ref": "#/components/schemas/CveIdentifier"
						},
						"type": "array"
					},
					"threatType": {
						"type": "string"
					},
					"threatvector": {
						"items": {},
						"type": "array"
					},
					"threatAffects": {
						"items": {
							"$ref": "#/components/schemas/ThreatAffect"
						},
						"type": "array"
					},
					"threatUpdates": {
						"items": {},
						"type": "array"
					},
					"remediationSteps": {
						"items": {
							"$ref": "#/components/schemas/RemediationStep"
						},
						"type": "array"
					},
					"ncscLink": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"versionedNode": {
						"type": "boolean"
					},
					"remediationIntro": {
						"type": "string"
					},
					"summary": {
						"type": "string"
					}
				},
				"required": [
					"severity",
					"category",
					"basePath",
					"sections",
					"threatId",
					"shortsummary",
					"publishedDate",
					"fullTaxonomyList",
					"seosummaryJson",
					"sourceOfThreatUpdates",
					"cyberAcknowledgements",
					"severityStatusHistory",
					"indicatorsCompromise",
					"cveIdentifiers",
					"threatType",
					"threatvector",
					"threatAffects",
					"threatUpdates",
					"remediationSteps",
					"ncscLink",
					"title",
					"versionedNode",
					"remediationIntro",
					"summary"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Welcome4": {
				"properties": {
					"ONS_CHD_GEO_LISTINGS_KEY": {
						"type": "number",
						"format": "double"
					},
					"GEOGRAPHY_CODE": {
						"type": "string"
					},
					"GEOGRAPHY_NAME": {
						"type": "string"
					},
					"WELSH_GEOGRAPHY_NAME": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"SI_NUMBER": {
						"type": "string"
					},
					"SI_TITLE": {
						"type": "string"
					},
					"DATE_OF_OPERATION": {
						"type": "string",
						"format": "date-time"
					},
					"DATE_OF_TERMINATION": {
						"type": "string",
						"format": "date-time"
					},
					"PARENT_GEOGRAPHY_CODE": {
						"type": "string"
					},
					"ENTITY_CODE": {
						"type": "string"
					},
					"ENTITY_OWNER": {
						"type": "string"
					},
					"IS_CURRENT": {
						"type": "number",
						"format": "double"
					},
					"SYSTEM_CREATED_DATE": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"ONS_CHD_GEO_LISTINGS_KEY",
					"GEOGRAPHY_CODE",
					"GEOGRAPHY_NAME",
					"WELSH_GEOGRAPHY_NAME",
					"SI_NUMBER",
					"SI_TITLE",
					"DATE_OF_OPERATION",
					"DATE_OF_TERMINATION",
					"PARENT_GEOGRAPHY_CODE",
					"ENTITY_CODE",
					"ENTITY_OWNER",
					"IS_CURRENT",
					"SYSTEM_CREATED_DATE"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Search3": {
				"properties": {
					"score": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"score"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IdentifierElement3": {
				"properties": {
					"system": {
						"type": "string"
					},
					"value": {
						"type": "string"
					},
					"extension": {
						"items": {
							"$ref": "#/components/schemas/IdentifierExtension3"
						},
						"type": "array"
					}
				},
				"required": [
					"system",
					"value",
					"extension"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Resource3": {
				"properties": {
					"resourceType": {
						"type": "string"
					},
					"id": {
						"type": "string"
					},
					"identifier": {
						"items": {
							"$ref": "#/components/schemas/IdentifierElement3"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta3"
					},
					"name": {
						"items": {
							"$ref": "#/components/schemas/Name3"
						},
						"type": "array"
					},
					"gender": {
						"type": "string"
					},
					"birthDate": {
						"type": "string",
						"format": "date-time"
					},
					"multipleBirthInteger": {
						"type": "number",
						"format": "double"
					},
					"deceasedDateTime": {
						"type": "string",
						"format": "date-time"
					},
					"address": {
						"items": {
							"$ref": "#/components/schemas/Address3"
						},
						"type": "array"
					},
					"telecom": {
						"items": {
							"$ref": "#/components/schemas/ResourceTelecom3"
						},
						"type": "array"
					},
					"contact": {
						"items": {
							"$ref": "#/components/schemas/Contact3"
						},
						"type": "array"
					},
					"generalPractitioner": {
						"items": {
							"$ref": "#/components/schemas/GeneralPractitioner3"
						},
						"type": "array"
					},
					"extension": {
						"items": {
							"$ref": "#/components/schemas/ResourceExtension3"
						},
						"type": "array"
					}
				},
				"required": [
					"resourceType",
					"id",
					"identifier",
					"meta",
					"name",
					"gender",
					"birthDate",
					"multipleBirthInteger",
					"deceasedDateTime",
					"address",
					"telecom",
					"contact",
					"generalPractitioner",
					"extension"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Entry3": {
				"properties": {
					"fullUrl": {
						"type": "string"
					},
					"search": {
						"$ref": "#/components/schemas/Search3"
					},
					"resource": {
						"$ref": "#/components/schemas/Resource3"
					}
				},
				"required": [
					"fullUrl",
					"search",
					"resource"
				],
				"type": "object",
				"additionalProperties": false
			},
			"personalDemo": {
				"properties": {
					"resourceType": {
						"type": "string"
					},
					"type": {
						"type": "string"
					},
					"timestamp": {
						"type": "string",
						"format": "date-time"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"entry": {
						"items": {
							"$ref": "#/components/schemas/Entry3"
						},
						"type": "array"
					}
				},
				"required": [
					"resourceType",
					"type",
					"timestamp",
					"total",
					"entry"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"name": "Authorization",
				"in": "header",
				"bearerFormat": "JWT"
			},
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-API-KEY"
			}
		}
	},
	"info": {
		"title": "Government Digital Service APIs",
		"version": "v1",
		"description": "The Government Digital Service (GDS) provides 12 APIs to get access to UK Govt. Services",
		"contact": {
			"name": "Oraclo LTD",
			"email": "admin@orcalo.co.uk",
			"url": "https://orcalo.co.uk/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/v1/CyberAlerts/limited/{bool}": {
			"get": {
				"operationId": "GetCyberAlerts1",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Welcome"
								}
							}
						}
					}
				},
				"summary": "This returns a list of cyber alerts",
				"tags": [
					"Cyber Alerts"
				],
				"security": [],
				"parameters": [
					{
						"description": "Adding the _limited=boolean parameter returns only a limited subset of the most important fields",
						"in": "path",
						"name": "bool",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/v1/CyberAlerts/page/{pageNo}": {
			"get": {
				"operationId": "GetCyberAlerts2",
				"responses": {
					"200": {
						"description": "0",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Welcome"
								}
							}
						}
					}
				},
				"summary": "This returns a list of cyber alerts",
				"tags": [
					"Cyber Alerts"
				],
				"security": [],
				"parameters": [
					{
						"description": "When requesting a full return of cyber alerts, you can return any page of the alerts shown by adding the &page=x, where x is the number of the page to be returned",
						"in": "path",
						"name": "pageNo",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/CyberAlerts/thread/{threadId}": {
			"get": {
				"operationId": "GetCyberAlerts3",
				"responses": {
					"200": {
						"description": "0",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Welcome1"
								}
							}
						}
					}
				},
				"summary": "This returns a list of cyber alerts",
				"tags": [
					"Cyber Alerts"
				],
				"security": [],
				"parameters": [
					{
						"description": "To return a single cyber alert, you must specify a threat ID, normally formatted CC-3327",
						"in": "path",
						"name": "threadId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/dataRegister/dbo/table/ONS_CHD_GEO_LISTINGS/query/{pageNo}": {
			"get": {
				"operationId": "QueryPageNo",
				"responses": {
					"200": {
						"description": "0",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Welcome4"
								}
							}
						}
					}
				},
				"summary": "This returns a list of cyber alerts",
				"tags": [
					"Data Registers Service"
				],
				"security": [],
				"parameters": [
					{
						"description": "To return a single cyber alert, you must specify a threat ID, normally formatted XXXXX",
						"in": "path",
						"name": "pageNo",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/dataRegister/dbo/table/ONS_CHD_GEO_LISTINGS/query/download/{bool}": {
			"get": {
				"operationId": "QueryTableDownload",
				"responses": {
					"200": {
						"description": "0",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/Welcome4"
										},
										{
											"type": "string"
										}
									]
								}
							}
						}
					}
				},
				"summary": "This returns a list of cyber alerts",
				"tags": [
					"Data Registers Service"
				],
				"security": [],
				"parameters": [
					{
						"description": "To return a single cyber alert, you must specify a threat ID, normally formatted XXXXX",
						"in": "path",
						"name": "bool",
						"required": true,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/v1/dataRegister/dbo/table/ONS_CHD_GEO_LISTINGS/query/filter/{filter}": {
			"get": {
				"operationId": "FilterItems",
				"responses": {
					"200": {
						"description": "0",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Welcome4"
								}
							}
						}
					}
				},
				"summary": "This returns a list of cyber alerts",
				"tags": [
					"Data Registers Service"
				],
				"security": [],
				"parameters": [
					{
						"description": "Filter argument must be <b>all</b>, <b>latest</b> or <b>archived</b>",
						"in": "path",
						"name": "filter",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/personalDemoGraphics": {
			"get": {
				"operationId": "GetPatientData",
				"responses": {
					"200": {
						"description": "personalDemo",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/personalDemo"
								}
							}
						}
					}
				},
				"summary": "sdfafasdf",
				"tags": [
					"Personal Demographics Service"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "AuthToken",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "NHSD-Session-URID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "X-Request-ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "header",
						"name": "X-Correlation-ID",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}